Upstream absorbed this fork's kernel hid_consumer, shared keyboard mapper,
audio-stream rebind + race test, UAC codec, consumer-HID routing, symbol
exports (audio-decoder-module, cjson-module, platform libgcc/esp_log), and
app_event_emit_close; those all follow upstream now. Kept from the fork:
Waveshare P4 board, st7796 spi-mode property, AetherLink CDC-ACM + UART
backends, AetherChat, tactility-audio module (AudioExports), esp_lcd_st7796
esp32p4 rule, and the external Audiobook Player app setup.

+++ Drivers/tca8418-module/include/drivers/tca8418.h
@@ -16,11 +16,11 @@ struct Tca8418Config {
     uint8_t rows;
     uint8_t columns;
     bool reverse_columns;
-    const uint8_t* keymap_lc;
+    const uint32_t* keymap_lc;
     uint32_t keymap_lc_length;
-    const uint8_t* keymap_uc;
+    const uint32_t* keymap_uc;
     uint32_t keymap_uc_length;
-    const uint8_t* keymap_sy;
+    const uint32_t* keymap_sy;
     uint32_t keymap_sy_length;
     uint8_t shift_row;
     uint8_t shift_col;
